Fix misuse of {} initialization
Narrowing is supposed to be forbidden in {} initialization, but Chromium
doesn't appear to care upstream.

+++ chromium/base/task/thread_pool/job_task_source.cc
@@ -90,14 +90,14 @@ JobTaskSource::State::DecrementWorkerCountFromWorkerAcquire() {
   const size_t value_before_sub =
       value_.fetch_sub(kWorkerCountIncrement, std::memory_order_acquire);
   DCHECK((value_before_sub >> kWorkerCountBitOffset) > 0);
-  return {value_before_sub};
+  return {uint32_t(value_before_sub)};
 }
 
 JobTaskSource::State::Value
 JobTaskSource::State::IncrementWorkerCountFromJoiningThread() {
   size_t value_before_add =
       value_.fetch_add(kWorkerCountIncrement, std::memory_order_relaxed);
-  return {value_before_add};
+  return {uint32_t(value_before_add)};
 }
 
 JobTaskSource::State::Value
@@ -105,7 +105,7 @@ JobTaskSource::State::DecrementWorkerCountFromJoiningThread() {
   const size_t value_before_sub =
       value_.fetch_sub(kWorkerCountIncrement, std::memory_order_relaxed);
   DCHECK((value_before_sub >> kWorkerCountBitOffset) > 0);
-  return {value_before_sub};
+  return {uint32_t(value_before_sub)};
 }
 
 JobTaskSource::State::Value JobTaskSource::State::Load() const {
